third_party/WebKit/LayoutTests/typedcssom/computedstyle/width.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>

<div id="testElement"></div>

<script>

var computedStyleMap = getComputedStyleMap(testElement);

test(function() {
  testElement.style.width = '10px';

  var result = computedStyleMap.get('width');
  assert_true(result instanceof CSSSimpleLength);
  assert_equals(result.cssText, '10px');
}, "Getting a 10px width results in a CSSSimpleLength");

test(function() {
  testElement.style.width = '20px';

  var result = computedStyleMap.getAll('width');
  assert_equals(result.length, 1);
  assert_equals(result[0].cssText, '20px');
}, "getAll for width returns a single value");

</script>
